+++ app/Http/Controllers/APIPendientes.php
@@ -29,12 +29,14 @@ class APIPendientes extends Controller
     {
         // Nuevo pendiente
         $this->validate($request, [
-            'nombre' => 'required|string|max:60',
-            'cantidad' => 'integer|nullable',
-            'precio' => 'numeric|required',
-            'descuento' => 'numeric|nullable',
-            'semanas' => 'integer',
+            'nombre' => 'required|string|min:3|max:60',
+            'cantidad' => 'integer|nullable|min:1|max:999999',
+            'precio' => 'numeric|required|max:99999999',
+            'descuento' => 'numeric|nullable|min:1|max:999',
+            'semanas' => 'integer|min:0|max:8'
+
         ]);
+
         $pendiente = Pendiente::create([
             'nombre' => $request->nombre,
             'precio' => $request->precio,
@@ -47,8 +49,15 @@ class APIPendientes extends Controller
     }
     public function destroy(Request $request)
     {
+
+        $id = filter_var($request->id, FILTER_VALIDATE_INT);
+
+        if(!$id) {
+            echo json_encode("Algo salió mal :( ");
+            return;
+        }
         // Eliminar pendiente
-        $pendiente = Pendiente::find($request->id);
+        $pendiente = Pendiente::find($id);
         $resultado = $pendiente->delete();
 
         echo json_encode($resultado);
